==> code/2_cluster_annotate_all.Rmd <==
## Load libraries

```{r}
#Load Helper Functions
source("C:/Users/Christopher He/OneDrive/UCSF/Singh Lab/diapause-sc-multiome/code/helper_functions.R")

suppressPackageStartupMessages({
  library(scCustomize)
})

#Set Out Directories
outs_dir <- "../results_all/"
fig_dir <- "../results_all/figures"
```

## Load Seurat File

```{r}
so_filtered <- readRDS(paste0(outs_dir, "so_final_filtered.rds"))
```

## Split into Annotation groups

split into 5 groups: pre-dia (25som, hb), dev(dev1d), early dia (dia1d, dia2d), mid dia (dia6d), and late dia (dia1mo, dia3mo)

```{r}
#make meta.data column with annotation group
Idents(so_filtered) <- "timepoint"
so_filtered <- RenameIdents(so_filtered, c("hb" = "predia",
                                           "25som" = "predia",
                                           "dia1d" = "earlydia",
                                           "dia2d" = "earlydia",
                                           "dia6d" = "middia",
                                           "dia1mo" = "latedia",
                                           "dia3mo" = "latedia",
                                           "dev1d" = "dev"))
so_filtered$timepoint_group <- Idents(so_filtered)

so_predia <- subset(so_filtered, timepoint_group == "predia")
so_dev <- subset(so_filtered, timepoint_group == "dev")
so_earlydia <- subset(so_filtered, timepoint_group == "earlydia")
so_middia <- subset(so_filtered,  timepoint_group == "middia")
so_latedia <- subset(so_filtered, timepoint_group == "latedia")

saveRDS(so_predia, paste0(outs_dir, "markers_predia/so_predia.rds"))
saveRDS(so_dev, paste0(outs_dir, "markers_dev/so_dev.rds"))
saveRDS(so_earlydia, paste0(outs_dir, "markers_earlydia/so_earlydia.rds"))
saveRDS(so_middia, paste0(outs_dir, "markers_middia/so_middia.rds"))
saveRDS(so_latedia, paste0(outs_dir, "markers_latedia/so_latedia.rds"))
```

### Save subsets

```{r}
#seurat processing for all 3 groups
so_predia <- runSeuratRNAProcessing(so_predia)
so_dev <- runSeuratRNAProcessing(so_dev)
so_earlydia <- runSeuratRNAProcessing(so_earlydia)
so_middia <- runSeuratRNAProcessing(so_middia, resolution = 0.6)
so_latedia <- runSeuratRNAProcessing(so_latedia)
```

## Pre-Annotation plots

### umap by cluster

```{r}
p <- DimPlot(so_predia, reduction = "umap", label = T, pt.size = 1) + NoLegend()
SaveFigure(p, path = "markers_predia/umap_clusters",
           height = 10, width = 10)

p <- DimPlot(so_dev, reduction = "umap", label = T, pt.size = 1) + NoLegend()
SaveFigure(p, path = "markers_dev/umap_clusters",
           height = 10, width = 10)

p <- DimPlot(so_earlydia, reduction = "umap", label = T, pt.size = 1) + NoLegend()
SaveFigure(p, path = "markers_earlydia/umap_clusters",
           height = 10, width = 10)

p <- DimPlot(so_middia, reduction = "umap", label = T, pt.size = 1) + NoLegend()
SaveFigure(p, path = "markers_middia/umap_clusters",
           height = 10, width = 10)

p <- DimPlot(so_latedia, reduction = "umap", label = T, pt.size = 1) + NoLegend()
SaveFigure(p, path = "markers_latedia/umap_clusters",
           height = 10, width = 10)
```

### umap by replicate

```{r}
p <- DimPlot(so_predia, group.by = "orig.ident", reduction = "umap", label = T, pt.size = 1)
SaveFigure(p, path = "markers_predia/umap_orig.ident",
           height = 10, width = 10)

p <- DimPlot(so_dev, group.by = "orig.ident", reduction = "umap", label = T, pt.size = 1)
SaveFigure(p, path = "markers_dev/umap_orig.ident",
           height = 10, width = 10)

p <- DimPlot(so_earlydia, group.by = "orig.ident", reduction = "umap", label = T, pt.size = 1)
SaveFigure(p, path = "markers_earlydia/umap_orig.ident",
           height = 10, width = 10)

p <- DimPlot(so_middia, group.by = "orig.ident", reduction = "umap", label = T, pt.size = 1)
SaveFigure(p, path = "markers_middia/umap_orig.ident",
           height = 10, width = 10)

p <- DimPlot(so_latedia, group.by = "orig.ident", reduction = "umap", label = T, pt.size = 1)
SaveFigure(p, path = "markers_latedia/umap_orig.ident",
           height = 10, width = 10)

```

### umap by cell cycle phase

requires cell cycle scoring first

```{r}
#cell cycle scoring
s.genes <- translate(cc.genes$s.genes, from = "Human", to = "N. furzeri (NCBI)")
g2m.genes <- translate(cc.genes$g2m.genes, from = "Human", to = "N. furzeri (NCBI)")

so_predia <- CellCycleScoring(so_predia, s.features = s.genes,
                                 g2m.features = g2m.genes)
so_dev <- CellCycleScoring(so_dev, s.features = s.genes,
                                 g2m.features = g2m.genes)
so_earlydia <- CellCycleScoring(so_earlydia, s.features = s.genes,
                                 g2m.features = g2m.genes)
so_middia <- CellCycleScoring(so_middia, s.features = s.genes,
                                 g2m.features = g2m.genes)
so_latedia <- CellCycleScoring(so_latedia, s.features = s.genes,
                                 g2m.features = g2m.genes)

p <- DimPlot(so_predia, reduction = "umap", group.by = "Phase", pt.size = 1)
SaveFigure(p, path = "markers_predia/umap_cellcycle",
           height = 10, width = 10)
p <- DimPlot(so_dev, reduction = "umap", group.by = "Phase", pt.size = 1)
SaveFigure(p, path ="markers_dev/umap_cellcycle",
           height = 10, width = 10)
p <- DimPlot(so_earlydia, reduction = "umap", group.by = "Phase", pt.size = 1)
SaveFigure(p, path = "markers_earlydia/umap_cellcycle",
           height = 10, width = 10)
p <- DimPlot(so_middia, reduction = "umap", group.by = "Phase", pt.size = 1)
SaveFigure(p, path = "markers_middia/umap_cellcycle",
           height = 10, width = 10)
p <- DimPlot(so_latedia, reduction = "umap", group.by = "Phase", pt.size = 1)
SaveFigure(p, path = "markers_latedia/umap_cellcycle",
           height = 10, width = 10)
```

## Marker Genes

### find all markers for each annotation gruop

```{r}
#find markers for all 5 groups
all_markers_predia <- FindAllMarkers(so_predia)
all_markers_dev <- FindAllMarkers(so_dev)
all_markers_earlydia <- FindAllMarkers(so_earlydia)
all_markers_middia <- FindAllMarkers(so_middia)
all_markers_latedia <- FindAllMarkers(so_latedia)
```

```{r}
#translate from NCBI annotation to our annotation and save
all_markers_predia$gene_name <- translate(all_markers_predia$gene)
all_markers_predia$gene_name <- paste0(all_markers_predia$gene,
                                  " (", all_markers_predia$gene_name, ")")

all_markers_dev$gene_name <- translate(all_markers_dev$gene)
all_markers_dev$gene_name <- paste0(all_markers_dev$gene,
                                  " (", all_markers_dev$gene_name, ")")

all_markers_earlydia$gene_name <- translate(all_markers_earlydia$gene)
all_markers_earlydia$gene_name <- paste0(all_markers_earlydia$gene,
                                  " (", all_markers_earlydia$gene_name, ")")

all_markers_middia$gene_name <- translate(all_markers_middia$gene)
all_markers_middia$gene_name <- paste0(all_markers_middia$gene,
                                  " (", all_markers_middia$gene_name, ")")

all_markers_latedia$gene_name <- translate(all_markers_latedia$gene)
all_markers_latedia$gene_name <- paste0(all_markers_latedia$gene,
                                  " (", all_markers_latedia$gene_name, ")")

write.csv(all_markers_predia,
          paste0(outs_dir, "markers_predia/all_markers_predia.csv"))
write.csv(all_markers_dev,
          paste0(outs_dir, "markers_dev/all_markers_dev.csv"))
write.csv(all_markers_earlydia,
          paste0(outs_dir, "markers_earlydia/all_markers_earlydia.csv"))
write.csv(all_markers_middia,
          paste0(outs_dir, "markers_middia/all_markers_middia.csv"))
write.csv(all_markers_latedia,
          paste0(outs_dir, "markers_latedia/all_markers_latedia.csv"))

```

### format top 20 markers for each cluster

```{r}
#format top 20 markers from each cluster for each group
top_markers_predia <- all_markers_predia %>%
  filter(avg_log2FC > 0.5) %>%
  group_by(cluster) %>%
  arrange(p_val_adj) %>%
  slice_head(n = 20) %>%
  mutate(rank = row_number()) %>%
  select(cluster, rank, gene_name) %>%
  pivot_wider(names_from = cluster, values_from = gene_name, names_prefix = "Cluster_", values_fill = list(gene_name = "")) %>%
  arrange(rank) %>%
  select(-rank)

top_markers_dev <- all_markers_dev %>%
  filter(avg_log2FC > 0.5) %>%
  group_by(cluster) %>%
  arrange(p_val_adj) %>%
  slice_head(n = 20) %>%
  mutate(rank = row_number()) %>%
  select(cluster, rank, gene_name) %>%
  pivot_wider(names_from = cluster, values_from = gene_name, names_prefix = "Cluster_", values_fill = list(gene_name = "")) %>%
  arrange(rank) %>%
  select(-rank)

top_markers_earlydia <- all_markers_earlydia %>%
  filter(avg_log2FC > 0.5) %>%
  group_by(cluster) %>%
  arrange(p_val_adj) %>%
  slice_head(n = 20) %>%
  mutate(rank = row_number()) %>%
  select(cluster, rank, gene_name) %>%
  pivot_wider(names_from = cluster, values_from = gene_name, names_prefix = "Cluster_", values_fill = list(gene_name = "")) %>%
  arrange(rank) %>%
  select(-rank)

top_markers_middia <- all_markers_middia %>%
  filter(avg_log2FC > 0.5) %>%
  group_by(cluster) %>%
  arrange(p_val_adj) %>%
  slice_head(n = 20) %>%
  mutate(rank = row_number()) %>%
  select(cluster, rank, gene_name) %>%
  pivot_wider(names_from = cluster, values_from = gene_name, names_prefix = "Cluster_", values_fill = list(gene_name = "")) %>%
  arrange(rank) %>%
  select(-rank)

top_markers_latedia <- all_markers_latedia %>%
  filter(avg_log2FC > 0.5) %>%
  group_by(cluster) %>%
  arrange(p_val_adj) %>%
  slice_head(n = 20) %>%
  mutate(rank = row_number()) %>%
  select(cluster, rank, gene_name) %>%
  pivot_wider(names_from = cluster, values_from = gene_name, names_prefix = "Cluster_", values_fill = list(gene_name = "")) %>%
  arrange(rank) %>%
  select(-rank)


write.csv(top_markers_predia,
          paste0(outs_dir, "markers_predia/top_markers_predia.csv"))
write.csv(top_markers_dev,
          paste0(outs_dir, "markers_dev/top_markers_dev.csv"))
write.csv(top_markers_earlydia,
          paste0(outs_dir, "markers_earlydia/top_markers_earlydia.csv"))
write.csv(top_markers_middia,
          paste0(outs_dir, "markers_middia/top_markers_middia.csv"))
write.csv(top_markers_latedia,
          paste0(outs_dir, "markers_latedia/top_markers_latedia.csv"))
```

## Cell Cycle Scoring

```{r}
s.genes <- translate(cc.genes$s.genes, from = "Human", to = "N. furzeri (NCBI)")
g2m.genes <- translate(cc.genes$g2m.genes, from = "Human", to = "N. furzeri (NCBI)")

so_dev_named <- CellCycleScoring(so_dev_named, s.features = s.genes,
                                 g2m.features = g2m.genes)

so_latedia_named <- CellCycleScoring(so_late_dia_named, s.features = s.genes,
                                 g2m.features = g2m.genes)

so_earlydia <- CellCycleScoring(so_earlydia, s.features = s.genes,
                                 g2m.features = g2m.genes)

so_predia <- CellCycleScoring(so_predia, s.features = s.genes,
                                 g2m.features = g2m.genes)



p <- DimPlot(so_dev_named, group.by = "Phase")
SaveFigure(p, path = "markers_dev/umap_cellcycle",
           height = 6, width = 6)
p <- DimPlot(so_latedia_named, group.by = "Phase")
SaveFigure(p, path = "markers_latedia/umap_cellcycle",
           height = 6, width = 6)
p <- DimPlot(so_earlydia_named, group.by = "Phase")
SaveFigure(p, path = "markers_earlydia/umap_cellcycle",
           height = 6, width = 6)
p <- DimPlot(so_predia_named, group.by = "Phase")
SaveFigure(p, path = "markers_predia/umap_cellcycle",
           height = 6, width = 6)
```


```{r}
all_markers <- FindAllMarkers(so_filtered)

top_markers <- all_markers %>%
  group_by(cluster) %>%
  arrange(p_val_adj) %>%
  slice_head(n = 10) %>%
  mutate(rank = row_number()) %>%
  select(cluster, rank, gene) %>%
  pivot_wider(names_from = cluster, values_from = gene, names_prefix = "Cluster_", values_fill = list(gene = "")) %>%
  arrange(rank) %>%
  select(-rank)

write.csv(all_markers, paste0(outs_dir, "/all_markers"))
write.csv(top_markers, paste0(outs_dir, "/top_markers"))
```

## Annotate Timepoint Groups

```{r}
#predia
cluster.ids <- c(
  'neuron_progenitor',
  'spinal_cord',
  'somite',
  'myotome/muscle_1',
  'telencephalon/optic_cup',
  'dividing_neuron',
  'epidermal',
  'head_kidney',
  'hibernating_unknown',
  'hatching_gland',
  'midbrain_hindbrain_pax2/5_en1/2',
  'differentiating_neuron',
  'endothelial',
  'periderm',
  'tail_bud',
  'neuron',
  'primitive_gut',
  'blood_progenitor',
  'myotome/muscle_2',
  'otic_vesicle',
  'pronephros',
  'fin_bud',
  'notochord',
  'neuron_dscam',
  'midbrain_hindbrain_wnt1/lmx1b'
)

names(cluster.ids) <- levels(so_predia)
so_predia_named <- RenameIdents(so_predia, cluster.ids)


p <- DimPlot(so_predia_named, reduction = "umap", label = T, repel = T, pt.size = 1) + NoLegend()
SaveFigure(p, path = "markers_predia/umap_annotated",
           height = 10, width = 10)

```

```{r}
#early dia
cluster.ids <- c(
  'spinal_cord',
  'neuron_progenitor',
  'somite',
  'telencephalon/optic_cup',
  'myotome/muscle',
  'differentiating_neuron',
  'floor_plate',
  'midbrain_hindbrain_pax2/5_en1/2',
  'epidermal',
  'hatching_gland',
  'endothelial',
  'head_kidney',
  'dividing_neuron',
  'periderm',
  'blood_progenitor',
  'otic_vesicle',
  'neuron_dscam',
  'neuron_sema3c',
  'fin_bud',
  'primitive_gut',
  'primitive_heart'
)

names(cluster.ids) <- levels(so_earlydia)
so_earlydia_named <- RenameIdents(so_earlydia, cluster.ids)


p <- DimPlot(so_earlydia_named, reduction = "umap", label = T, repel = T, pt.size = 1) + NoLegend()

SaveFigure(p, path = "markers_earlydia/umap_annotated",
           height = 10, width = 10)
```

```{r}
#mid dia
cluster.ids <- c(
  'neuron_progenitor',
  'hibernating_neuron_progenitor',
  'spinal_cord',
  'epidermal',
  'differentiating_neuron',
  'somite',
  'hindbrain',
  'telencephalon/optic_cup',
  'myotome/muscle',
  'hatching_gland',
  'floor_plate',
  'blood',
  'midbrain_hindbrain_pax2/5_en1/2',
  'endothelial',
  'periderm',
  'epidermal_lye',
  'neuron_sema3c'
)

names(cluster.ids) <- levels(so_middia)
so_middia_named <- RenameIdents(so_middia, cluster.ids)


p <- DimPlot(so_middia_named, reduction = "umap", label = T, repel = T, pt.size = 1) + NoLegend()
SaveFigure(p, path = "markers_middia/umap_annotated",
           height = 10, width = 10)
```

```{r}
#late dia
cluster.ids <- c(
  'neuron_progenitor',
  'hibernating_neuron_progenitor_1',
  'hibernating_neuron_progenitor_2',
  'somite',
  'midbrain_hindbrain_wnt1/lmx1b',
  'spinal_cord',
  'epidermal',
  'neuron',
  'midbrain_hindbrain_pax2/5_en1/2',
  'hatching_gland',
  'endothelial',
  'telencephalon/optic_cup',
  'blood',
  'myotome/muscle',
  'immune_cell'
)

names(cluster.ids) <- levels(so_latedia)
so_latedia_named <- RenameIdents(so_latedia, cluster.ids)


p <- DimPlot(so_latedia_named, reduction = "umap", label = T, repel = T, pt.size = 1) + NoLegend()
SaveFigure(p, path = "markers_latedia/umap_annotated",
           height = 10, width = 10)
```

```{r}
#dev
cluster.ids <- c(
  'neuron_progenitor_1',
  'neuron_progenitor_2',
  'somite',
  'neuron',
  'spinal_cord',
  'differentiating_neuron_1',
  'epidermal_1',
  'myotome/muscle',
  'telencephalon/optic_cup',
  'floor_plate',
  'midbrain_hindbrain_pax2/5_en1/2',
  'periderm',
  'head_kidney',
  'otic_vesicle',
  'endothelial',
  'hatching_gland',
  'neural_crest',
  'neuron_neurod1',
  'midbrain_hindbrain_wnt1/lmx1b',
  'pigment_cell',
  'blood',
  'neuron_sema3c',
  'neuron_dscam',
  'notochord',
  'epidermal_unknown',
  'pronephros',
  'differentiating_neuron_2',
  'primitive_gut'
)

names(cluster.ids) <- levels(so_dev)
so_dev_named <- RenameIdents(so_dev, cluster.ids)

so_dev_named$annotation <- Idents(so_dev_named)

p <- DimPlot(so_dev_named, reduction = "umap", label = T, repel = T, pt.size = 1) + NoLegend()
SaveFigure(p, path = "markers_dev/umap_annotated",
           height = 10, width = 10)
```

```{r}
#save individual objects
saveRDS(so_predia_named, paste0(outs_dir, "markers_predia/so_predia.rds"))
saveRDS(so_dev_named, paste0(outs_dir, "markers_dev/so_dev.rds"))
saveRDS(so_earlydia_named, paste0(outs_dir, "markers_earlydia/so_earlydia.rds"))
saveRDS(so_middia_named, paste0(outs_dir, "markers_middia/so_middia.rds"))
saveRDS(so_latedia_named, paste0(outs_dir, "markers_latedia/so_latedia.rds"))
```

## Add annotations to full object

```{r}
#get annotations for each subset
predia <- Idents(so_predia_named)
earlydia <- Idents(so_earlydia_named)
middia <- Idents(so_middia_named)
latedia <- Idents(so_latedia_named)
dev <- Idents(so_dev_named)

#combine annotations
annotations_merged <- data.frame(c(predia, earlydia, middia, latedia, dev)) #merged annotation
annotations_merged <- annotations_merged[row.names(so_filtered@meta.data),]

#add to full seurat object
so_filtered$annotation_merged <- annotations_merged

#add column appending group
so_filtered$annotation_grouped <- paste0(so_filtered$annotation_merged, "_", so_filtered$timepoint_group)

#add simplified annotation by comining similar cell types
#remove "_number" suffix or "?" prefix
so_filtered$annotation_merged <- gsub("^\\?|_\\d+", "", so_filtered$annotation_merged)

#make plots
p <- DimPlot(so_filtered, group.by = "annotation_merged", reduction = "umap", repel = T, label = T) + NoLegend()
SaveFigure(p, "figures/umap_annotated_combined", height = 10, width = 10)

```

### Save Annotated object

```{r}
#save object

saveRDS(so_filtered, paste0(outs_dir, "so_annotated.rds"))
# so_filtered <- readRDS(paste0(outs_dir, "so_annotated.rds"))
```

## Broad cell types for nicer visualization

```{r}
broad <- c(
  "myotome/muscle" = "myotome/muscle",
  "dividing_neuron" = "neuron_progenitor",
  "hibernating_neuron_progenitor" = "neuron_progenitor",
  "periderm" = "periderm",
  "spinal_cord" = "spinal_cord",
  "midbrain_hindbrain_pax2/5_en1/2" = "midbrain_hindbrain_pax2/5_en1/2",
  "neuron_progenitor" = "neuron_progenitor",
  "somite" = "somite",
  "neuron" = "neuron",
  "tail_bud" = "tail_bud",
  "differentiating_neuron" = "differentiating_neuron",
  "hatching_gland" = "hatching_gland",
  "pigment_cell" = "pigment_cell",
  "epidermal" = "epidermal",
  "hindbrain" = "neuron_progenitor",
  "floor_plate" = "neuron_progenitor",
  "telencephalon/optic_cup" = "telencephalon/optic_cup",
  "endothelial" = "endothelial",
  "head_kidney" = "head_kidney",
  "blood" = "blood",
  "otic_vesicle" = "otic_vesicle",
  "hibernating_unknown" = "hibernating_unknown",
  "neuron_neurod1" = "neuron",
  "neuron_sema3c" = "neuron_sema3c",
  "neuron_dscam" = "neuron_dscam",
  "midbrain_hindbrain_wnt1/lmx1b" = "midbrain_hindbrain_wnt1/lmx1b",
  "neural_crest" = "neural_crest",
  "epidermal_unknown" = "epidermal",
  "blood_progenitor" = "blood",
  "primitive_gut" = "primitive_gut",
  "pronephros" = "pronephros",
  "primitive_heart" = "primitive_heart",
  "fin_bud" = "fin_bud",
  "notochord" = "notochord",
  "epidermal_lye" = "epidermal",
  "immune_cell" = "immune_cell"
)

Idents(so_filtered) <- "annotation_merged"
so_filtered <- RenameIdents(so_filtered, broad)
so_filtered$annotation_broad <- Idents(so_filtered)
```


```{r}
p <- DimPlot(so_filtered, reduction = "umap", cols = colors_broad, pt.size = 0.7)

SaveFigure(p, "figures/umap_annotated_broad", height = 10, width = 17)

```

```{r}
Idents(so_filtered) <- "annotation_grouped"
Cluster_Highlight_Plot(seurat_object = so_filtered, cluster_name = "periderm_latedia", highlight_color = "gold",
background_color = "lightgray",  pt.size = 1, reduction = "umap")

Cluster_Highlight_Plot(seurat_object = so_filtered, cluster_name = "differentiating_neuron_middia", highlight_color = "gold",
background_color = "lightgray",  pt.size = 1, reduction = "umap")

Cluster_Highlight_Plot(seurat_object = so_filtered, cluster_name = "differentiating_neuron_earlydia", highlight_color = "gold",
background_color = "lightgray",  pt.size = 1, reduction = "umap")

Cluster_Highlight_Plot(seurat_object = so_filtered, cluster_name = "differentiating_neuron_predia", highlight_color = "gold",
background_color = "lightgray",  pt.size = 1, reduction = "umap")

Cluster_Highlight_Plot(seurat_object = so_filtered, cluster_name = "neuron_predia", highlight_color = "gold",
background_color = "lightgray",  pt.size = 1, reduction = "umap")
```

## Neuronal Subclustering

```{r}
so_neurons <- subset(
  x = so_filtered,
  annotation_combined %in% c(
    'dividing_neuron',
    'hibernating_unknown',
    'spinal_cord',
    'midbrain_hindbrain_boundary',
    'neuron_progenitor',
    'neuron',
    'hibernating_neuron_progenitor',
    'differentiating_neuron',
    'telencephalon/optic_cup',
    'trigeminal_ganglion?',
    'floor_plate_neural_tube',
    'hindbrain'
  )
)

so_neurons <- runSeuratRNAProcessing(so_neurons)

p <- DimPlot(so_neurons, reduction = "umap", group.by = "annotation_combined", label = T, repel = T) + NoLegend()

SaveFigure(p, "figures/umap_neurons", height = 7, width = 7)
```

```{r}
annotation_broad_neuron <- c(
  "dividing_neuron" = "neuron_progenitor",
  "hibernating_unknown" = "neuron_progenitor",
  "spinal_cord" = "spinal_cord",
  "midbrain_hindbrain_boundary" = "midbrain_hindbrain_boundary",
  "neuron_progenitor" = "neuron_progenitor",
  "neuron" = "neuron",
  "hibernating_neuron_progenitor" = "neuron_progenitor",
  "differentiating_neuron" = "differentiating_neuron",
  "telencephalon/optic_cup" = "telencephalon/optic_cup",
  "trigeminal_ganglion?" = "neuron",
  "floor_plate_neural_tube" = "neuron_progenitor",
  "hindbrain" = "neuron_progenitor"
)

Idents(so_neurons) <- "annotation_neuron_broad"
so_neurons <- RenameIdents(so_neurons, annotation_broad_neuron)
so_neurons$annotation_broad <- Idents(so_neurons)

so_neurons$annotation_broad_devtraj <- paste0(so_neurons$annotation_broad, "_", so_neurons$devtraj)



DimPlot(so_neurons, label = T, repel = T, reduction = "umap")
```

```{r}
broad_cols_neuron <- c(
  "neuron_progenitor_dia" = "#1F78B4",
  "spinal_cord_dia" = "#33A02C",
  "midbrain_hindbrain_boundary_dia" = "#E31A1C",
  "telencephalon/optic_cup_dia" = "#FF7F00",
  "differentiating_neuron_dia" = "#6A3D9A",
  "neuron_dia" = "#B15928",
  "neuron_progenitor_pre-dia" = "#A6CEE3",
  "spinal_cord_pre-dia" = "#B2DF8A",
  "midbrain_hindbrain_boundary_pre-dia" = "#FB9A99",
  "telencephalon/optic_cup_pre-dia" = "#FDBF6F",
  "differentiating_neuron_pre-dia" = "#CAB2D6",
  "neuron_pre-dia" = "#FFFF99",
  "neuron_progenitor_dev" = "#A6CEE3",
  "spinal_cord_dev" = "#B2DF8A",
  "midbrain_hindbrain_boundary_dev" = "#FB9A99",
  "telencephalon/optic_cup_dev" = "#FDBF6F",
  "differentiating_neuron_dev" = "#CAB2D6",
  "neuron_dev" = "#FFFF99"
)

p <- DimPlot(so_neurons, label = T, repel = T, reduction = "umap",
             group.by = "annotation_neuron_broad", cols = broad_cols_neuron)
SaveFigure(p, "")

```
