==> code/8_rna_velocity.Rmd <==
---
title: "8_rna_velocity"
---
```{r}
#import helper functions
source("C:/Users/Christopher He/OneDrive/UCSF/Singh Lab/diapause-sc-multiome/code/helper_functions.R")

suppressPackageStartupMessages({
  library(SeuratWrappers)
})

#set output and figure directories
outs_dir <- "../results_all/lineage_subsets/"
fig_dir <- "../results_all/figures/"
```


# Write files for conversion to adata
- lineages: neuron, epidermal, somite, hatching gland, endothelial, blood
```{r}
#read in subsetted seurat objects
so_neuron <- readRDS("../results_all/lineage_subsets/so_neurons.rds")
so_epidermal <- readRDS("../results_all/lineage_subsets/so_epidermal.rds")
so_somite <- readRDS("../results_all/lineage_subsets/so_somite.rds")
so_hatch <- readRDS("../results_all/lineage_subsets/so_hatch.rds")
so_endothelial <- readRDS("../results_all/lineage_subsets/so_endothelial.rds")
so_blood <- readRDS("../results_all/lineage_subsets/so_blood.rds")

so_list <- c(so_neuron, so_epidermal, so_somite, so_hatch, so_endothelial, so_blood)


#First, plot RNA, ATAC, and wnn umaps by timepoint and cell type
for(i in 1:6){
  so <- so_list[[i]]
  p1 <- DimPlot(so, reduction = "umap.rna", group.by = "timepoint_group", cols = colors_timepoints)
  p2 <- DimPlot(so, reduction = "umap.rna", group.by = "annotation_broad", cols = colors_broad, label = T)
  p3 <- DimPlot(so, reduction = "umap.rna", group.by = "RNA_snn_res.0.8", label = T)

  
  p4 <- DimPlot(so, reduction = "umap.atac", group.by = "timepoint_group", cols = colors_timepoints)
  p5 <- DimPlot(so, reduction = "umap.atac", group.by = "annotation_broad", cols = colors_broad, label = T)
  p6 <- DimPlot(so, reduction = "umap.atac", group.by = "ATAC_snn_res.0.8", label = T)

  p7 <- DimPlot(so, reduction = "wnn.umap", group.by = "timepoint_group", cols = colors_timepoints)
  p8 <- DimPlot(so, reduction = "wnn.umap", group.by = "annotation_broad", cols = colors_broad, label = T)
  
  SaveFigure((p1|p2|p3)/(p4|p5|p6)/(p7|p8) & NoLegend() & NoAxes(), paste0(i, "_subset_umaps"), width = 10, height = 10)
}
```


```{r}
#write files for adata conversion
for(i in 1:6){
  so <- so_list[[i]]
  
  #save barcodes to metadata
  so$barcode <- colnames(so)
  
  #save umap coordinates to metadata
  so$umap.rna_1 <- so@reductions$umap.rna@cell.embeddings[,1]
  so$umap.rna_2 <- so@reductions$umap.rna@cell.embeddings[,2]
  so$umap.atac_1 <- so@reductions$umap.atac@cell.embeddings[,1]
  so$umap.atac_2 <- so@reductions$umap.atac@cell.embeddings[,2]
  so$umap.wnn_1 <- so@reductions$wnn.umap@cell.embeddings[,1]
  so$umap.wnn_2 <- so@reductions$wnn.umap@cell.embeddings[,2]
  
  # write metadata to csv
  write.csv(so@meta.data, file=paste0(outs_dir, i, '_metadata.csv'), quote=F, row.names=F)
  
  # write pca matrix to csv
  write.csv(so@reductions$pca@cell.embeddings, file=paste0(outs_dir, i, '_pca.csv'), quote=F, row.names=F)
  
  # extract neighborhood graph
  nn_idx <- so@neighbors$weighted.nn@nn.idx
  nn_dist <- so@neighbors$weighted.nn@nn.dist
  nn_cells <- so@neighbors$weighted.nn@cell.names
  
  # save neighborhood graph
  write.table(nn_idx, paste0(outs_dir, i, "_nn_idx.txt"), sep = ',', row.names = F, col.names = F, quote = F)
  write.table(nn_dist, paste0(outs_dir, i, "_nn_dist.txt"), sep = ',', row.names = F, col.names = F, quote = F)
  write.table(nn_cells, paste0(outs_dir, i, "_nn_cells.txt"), sep = ',', row.names = F, col.names = F, quote = F)
}
```

# Activate Python environnment
Note: Using python version through reticulate. scVelo and multiVelo were installed in the multiome-velocyto conda environment
```{r}
reticulate::use_condaenv("multiome-velocyto")
library(reticulate)
```

```{python}
import scipy
import os
import numpy as np
import pandas as pd
import scanpy as sc
import scvelo as scv
import multivelo as mv
import matplotlib.pyplot as plt
from pathlib import Path
```

```{python}
sc.settings.cachedir = Path('../results_all/cache')
scv.settings.figdir = '../results_all/lineage_subsets'
scv.settings.plot_prefix = ""
scv.settings.verbosity = 3
scv.settings.presenter_view = True
scv.set_figure_params('scvelo')
pd.set_option('display.max_columns', 100)
pd.set_option('display.max_rows', 200)
np.set_printoptions(suppress=True)
```

# Process adata object (wynton)
```{python}
#Read in loom file (contains spliced and unspliced counts)
adata_rna = sc.read("aggr_batch1-3/velocyto/velocyto_combined.loom", cache=True)

#rename barcodes to seurat format ("[barcode]-n" where n correspondes to sample)
suffixes = sum([[f"-{i}"] * n for i, n in enumerate([10177, 9774, 10510, 6575, 10736, 5993, 17082, 12245, 7928, 12501, 20000, 12563], start=1)], [])
adata_rna.obs_names = [x.split(':')[1][:-1] for x in adata_rna.obs_names]
adata_rna.obs_names = [a + b for a, b in zip(adata_rna.obs_names, suffixes)] #append "-n" suffix to each barcode

#read in metadata and pca table from seurat object
cell_meta = pd.read_csv("results_all/lineage_subsets/" + prefix + "_metadata.csv")

#filter barcodes in seurat object
barcodes = cell_meta.barcode
adata_rna = adata_rna[adata_rna.obs_names.isin(barcodes)].copy()

#reorder metadata in order of adata
cell_meta = cell_meta.iloc[pd.Index(cell_meta['barcode']).get_indexer(adata_rna.obs_names)] 

#read in pca
pca = pd.read_csv("results_all/lineage_subsets/" + prefix + "_pca.csv")
pca = pca.loc[cell_meta.index] # reorder pca in order of adata

#add metadata to adata
cell_meta = cell_meta.set_index('barcode')
adata_rna.obs = cell_meta

#set pca and umap
adata_rna.obsm['X_pca'] = pca.to_numpy()
adata_rna.obsm['X_umap'] = np.vstack((adata_rna.obs['umap.rna_1'].to_numpy(), adata_rna.obs['umap.rna_2'].to_numpy())).T

#Top 2000 variable genes are used for downstream analyses.
scv.pp.filter_and_normalize(adata_rna, min_shared_counts=10, n_top_genes=2000)

#read and preprocess atac counts
adata_atac = sc.read_10x_mtx('aggr_batch1-3/filtered_feature_bc_matrix/', var_names='gene_symbols', cache=True, gex_only=False)
adata_atac = adata_atac[:,adata_atac.var['feature_types'] == "Peaks"]

adata_atac = mv.aggregate_peaks_10x(adata_atac,
                                    'aggr_batch1-3/atac_peak_annotation.tsv',
                                    'aggr_batch1-3/feature_linkage.bedpe')
#barcodes from seurat object
adata_atac = adata_atac[adata_atac.obs_names.isin(barcodes)].copy()
# We normalize aggregated peaks with TF-IDF.
mv.tfidf_norm(adata_atac)


#Finding shared barcodes and features between RNA and ATAC
shared_genes = pd.Index(np.intersect1d(adata_rna.var_names, adata_atac.var_names))
shared_cells = pd.Index(np.intersect1d(adata_rna.obs_names, adata_atac.obs_names))
adata_rna = adata_rna[shared_cells, shared_genes]
adata_atac = adata_atac[shared_cells, shared_genes]


#scVelo preprocessing steps
scv.pp.normalize_per_cell(adata_rna)
scv.pp.log1p(adata_rna)
scv.pp.moments(adata_rna, n_pcs=30, n_neighbors=50)

#calculate velocity using scVelo
scv.tl.velocity(adata_rna, mode='stochastic')
scv.tl.velocity_graph(adata_rna)

# Smoothing gene aggregagted peaks by neighbors
nn_idx = np.loadtxt("results_all/lineage_subsets/" + prefix + "_nn_idx.txt", delimiter=',')
nn_dist = np.loadtxt("results_all/lineage_subsets/" + prefix + "_nn_dist.txt", delimiter=',')
nn_cells = pd.Index(pd.read_csv("results_all/lineage_subsets/" + prefix + "_nn_cells.txt", header=None)[0])

mv.knn_smooth_chrom(adata_atac, nn_idx, nn_dist)

# Running multi-omic dynamical model
mv.settings.VERBOSITY = 0
adata_result = mv.recover_dynamics_chrom(
  adata_rna,
 adata_atac,
 max_iter=5,
 init_mode="invert",
 parallel=True,
 save_plot=False,
 rna_only=False,
 fit=True,
 n_anchors=500
)
            
mv.velocity_graph(adata_result)
mv.latent_time(adata_result)


#save results             
adata_atac.write("results_all/lineage_subsets/" + prefix + "_adata_atac.h5ad")
adata_rna.write("results_all/lineage_subsets/" + prefix + "_adata_rna.h5ad")
adata_result.write("results_all/lineage_subsets/" + prefix + "_multivelo_result.h5ad")
```

# Read in results from wynton
```{python}
# adata_result = sc.read_h5ad("../results_all/lineage_subsets/1_multivelo_result.h5ad")
# adata_rna = sc.read_h5ad("../results_all/velocity/adata_rna.h5ad")
# adata_atac = sc.read_h5ad("../results_all/velocity/adata_atac.h5ad")
```

# ScVelo
```{python}
#iterate through all lineages, compute velocity, pseudotime, etc. and plot, save adata object

for j in range(1,7):
  i = str(j)
  adata_rna = sc.read_h5ad("../results_all/lineage_subsets/" + i + "_adata_rna.h5ad")
  
  #graph velocity
  scv.pl.velocity_embedding_grid(adata_rna, basis='umap', color='timepoint_group', save= i + '_embedding_grid.png', title='', scale=0.25, palette = ['#FD8D3C','#C6DBEF', '#2171B5', '#6BAED6', '#800080'])
  scv.pl.velocity_embedding_stream(adata_rna, basis='umap', color='timepoint_group', save= i + '_embedding_stream.png', title='', palette = ['#FD8D3C','#C6DBEF', '#2171B5', '#6BAED6', '#800080'])
  
  #top velocity genes
  scv.tl.rank_velocity_genes(adata_rna, groupby='timepoint_group', min_corr=.3)
  df = pd.DataFrame(adata_rna.uns['rank_velocity_genes']['names'])
  df.to_csv("../results_all/lineage_subsets/" + i + "_velocity_genes.csv")
  
  #velocity graph and pseudotime
  scv.pl.velocity_graph(adata_rna, threshold=.5, color = 'timepoint_group', save= i + '_velocity_graph.png')
  scv.tl.velocity_pseudotime(adata_rna)
  scv.pl.scatter(adata_rna, color='velocity_pseudotime', cmap='gnuplot', save = i + '_velocity_pseudotime.png')
  
  hm = scv.pl.heatmap(adata_rna, var_names=adata_rna.var.index, sortby='velocity_pseudotime', col_color='timepoint_group', n_convolve=100, save = i + '_pseudotime_heatmap.png', show = False, figsize = (8,10))
  gene_order = hm.data2d.index.tolist()
  pd.DataFrame(gene_order).to_csv("../results_all/lineage_subsets/" + i + "_heatmap_gene_order.csv", sep = "\n", index = False)
  
  adata_rna.write("../results_all/lineage_subsets/" + i + "_adata_rna.h5ad", compression='gzip')

```

## Splicing proportions
```{python}
scv.pl.proportions(adata_rna, groupby='annotation_broad')
scv.pl.proportions(adata_rna, groupby='timepoint_group')
```

## Compute Velocities
```{python}
#compute velocity for each gene
scv.tl.velocity(adata_rna, mode='stochastic')

#compute transition probabilities using gene velocities
scv.tl.velocity_graph(adata_rna)
```

## Visualize velocity fields
```{python}
#visualize velocity fields
scv.pl.velocity_embedding_grid(adata_rna, basis='umap', color='annotation_broad', save='embedding_grid.pdf', title='', scale=0.25)
scv.pl.velocity_embedding_stream(adata_rna, basis='umap', color='annotation_broad', save='embedding_stream.pdf', title='')


scv.pl.velocity_embedding_grid(adata_rna, basis='umap', color='timepoint_group', save='embedding_grid.pdf', title='', scale=0.25, palette = ['#FD8D3C','#C6DBEF', '#2171B5', '#6BAED6', '#800080'])
scv.pl.velocity_embedding_stream(adata_rna, basis='umap', color='timepoint_group', save='embedding_stream.pdf', title='', palette = ['#FD8D3C','#C6DBEF', '#2171B5', '#6BAED6', '#800080'])
```

## Plot some genes
```{python}
# plot velocity of a selected gene
scv.pl.velocity(adata_rna, var_names=['jade2'], color='timepoint_group')
scv.pl.velocity(adata_rna, var_names=['elavl3'], color='annotation_broad')

```

## Identify Important genes
```{python}
scv.tl.rank_velocity_genes(adata_rna, groupby='timepoint_group', min_corr=.3)

df = pd.DataFrame(adata_rna.uns['rank_velocity_genes']['names'])
df.head()
```


```{python}
kwargs = dict(frameon=False, size=10, linewidth=1.5, color = 'timepoint_group')

scv.pl.scatter(adata_rna, df['latedia'][:5], ylabel='latedia', dpi = 500, save='latedia_neuron_genes.pdf', **kwargs)
scv.pl.scatter(adata_rna, df['earlydia'][:5], ylabel='earlydia', save='earlydia_neuron_genes.png', **kwargs)
```

## Speed and coherence
Two more useful stats: - The speed or rate of differentiation is given by the length of the velocity vector. - The coherence of the vector field (i.e., how a velocity vector correlates with its neighboring velocities) provides a measure of confidence.
```{python}
scv.tl.velocity_confidence(adata_rna)
keys = 'velocity_length', 'velocity_confidence'
scv.pl.scatter(adata_rna, c=keys, cmap='coolwarm', perc=[5, 95])
```

## Velocity graph and pseudotime
```{python}
scv.pl.velocity_graph(adata_rna, threshold=.5, color = 'timepoint_group')
```

```{python}
scv.tl.velocity_pseudotime(adata_rna)
scv.pl.scatter(adata_rna, color='velocity_pseudotime', cmap='gnuplot')

scv.pl.heatmap(adata_rna, var_names=adata_rna.var.index, sortby='velocity_pseudotime', col_color='timepoint_group', n_convolve=100, save = 'pseudotime_heatmap_neuron.png')

```
## PAGA velocity graph
```{python}
!pip install python-igraph --upgrade --quiet
```


```{python}
# this is needed due to a current bug - bugfix is coming soon.
adata_rna.uns['neighbors']['distances'] = adata_rna.obsp['distances']
adata_rna.uns['neighbors']['connectivities'] = adata_rna.obsp['connectivities']

scv.tl.paga(adata_rna, groups='timepoint_group')
df = scv.get_df(adata_rna, 'paga/transitions_confidence', precision=2).T
df.style.background_gradient(cmap='Blues').format('{:.2g}')

df
```

```{python}
scv.pl.paga(adata_rna, basis='umap', size=50, alpha=.1, min_edge_width=2, node_size_scale=1.5)
```


# Multi-Velo
```{python}
#Read in results from wynton

adata_result = sc.read_h5ad("../results_all/velocity/multivelo_result.h5ad")
adata_rna = sc.read_h5ad("../results_all/velocity/adata_rna.h5ad")
adata_atac = sc.read_h5ad("../results_all/velocity/adata_atac.h5ad")
```


## Multi-velo plots
```{python}
mv.pie_summary(adata_result)
plt.show()

mv.switch_time_summary(adata_result)
plt.show()

mv.likelihood_plot(adata_result)
plt.show()

```

```{python}
mv.velocity_embedding_stream(adata_result, basis='umap', color = 'annotation_broad')
scv.pl.scatter(adata_result, color='latent_time', color_map='gnuplot', size=80)
```

# scVelo Dynamical model

```{python}
scv.pp.filter_and_normalize(adata_rna, min_shared_counts=10, n_top_genes=2000)
scv.pp.moments(adata_rna, n_pcs=30, n_neighbors=30)
```

```{python}
scv.tl.recover_dynamics(adata_rna)
```

```{python}
scv.tl.velocity(adata_rna, mode='dynamical')
scv.tl.velocity_graph(adata_rna)

adata_rna.write('../results_all/velocity/adata_rna_dynamical.h5ad', compression='gzip')
adata_rna = sc.read('../results_all/velocity/adata_rna_dynamical.h5ad')
```
```{python}
scv.pl.velocity_embedding_grid(adata_rna, basis='umap', color='annotation_broad', save='embedding_grid.pdf', title='', scale=0.25)

scv.pl.velocity_embedding_stream(adata_rna, basis='umap', color = "timepoint_group")
```


```{python}
scv.tl.latent_time(adata_rna)
scv.pl.scatter(adata_rna, color='latent_time', color_map='gnuplot', size=80)
scv.pl.scatter(adata_rna, color='velocity_pseudotime', color_map='gnuplot', size=80)

```

```{python}
top_genes = adata_rna.var['fit_likelihood'].sort_values(ascending=False).index[:300]
scv.pl.heatmap(adata_rna, var_names=top_genes, sortby='velocity_pseudotime', col_color='annotation_broad', n_convolve=100)
```
```{python}
adata_rna.obs
```
