==> code/9_lineage_subclustering.Rmd <==
---
title: "9_lineage_subclustering"
---

```{r}
#import helper functions
source("C:/Users/Christopher He/OneDrive/UCSF/Singh Lab/diapause-sc-multiome/code/helper_functions.R")

suppressPackageStartupMessages({
  library(monocle3)
  library(SeuratWrappers)
})




#set output and figure directories
outs_dir <- "../results_all/lineage_subsets/"
fig_dir <- "../results_all/lineage_subsets/"
```

```{r}
subset_process_RNA_ATAC <- function(so, rna.cts){
  #subset cells based on rna annotation
  rna.cells <- WhichCells(so, expression = annotation_broad %in% rna.cts)
  so_subset <- subset(so, cells = rna.cells)
  
  #process RNA
  DefaultAssay(so_subset) <- "RNA"
  so_subset <- NormalizeData(so_subset)
  so_subset <- FindVariableFeatures(so_subset)
  so_subset <- ScaleData(so_subset)
  so_subset <- RunPCA(so_subset, verbose = FALSE)
  so_subset <- FindNeighbors(object = so_subset, dims = 1:30)
  so_subset <- FindClusters(object = so_subset, resolution = 0.5)
  so_subset <- RunUMAP(so_subset, dims = 1:30, reduction.name = "umap_RNA")

  #process ATAC
  DefaultAssay(so_subset) <- "ATAC"
  so_subset <- RunTFIDF(so_subset)
  so_subset <- FindTopFeatures(so_subset, min.cutoff = 'q0')
  so_subset <- RunSVD(so_subset)
  so_subset <- FindNeighbors(object = so_subset, dims = 1:30)
  so_subset <- FindClusters(object = so_subset, resolution = 0.5)
  so_subset <- RunUMAP(so_subset, reduction = 'lsi', dims = 2:30, reduction.name = "umap_atac") # optional
  
  #WNN
  so_subset <- FindMultiModalNeighbors(so_subset, reduction.list = list("pca", "lsi"), dims.list = list(1:30, 2:30))
  so_subset <- RunUMAP(so_subset, nn.name = "weighted.nn", reduction.name = "umap_wnn") # optional

  return(so_subset)
}
```



```{r}
so <- readRDS("../results_all/so_wnn.rds")
```


```{r}
table(so$annotation_ATAC_broad)
```

```{r}
subset_list <- list(
  "1_neuron" = list("neuron", "neuron_progenitor", "differentiating_neuron"),
  "2_midbrain_hindbrain_pax_en" = list("midbrain_hindbrain_pax2/5_en1/2"),
  "3_telencephalon" = list("telencephalon/optic_cup"),
  "4_epidermal" = list("epidermal"),
  "5_periderm" = list("periderm"),
  "6_myotome" = list("myotome/muscle"),
  "7_somite" = list("somite"),
  "8_endothelial" = list("endothelial"),
  "9_blood" = list("blood"),
  "10_hatching_gland" = list("hatching_gland")
)
```

# create, process, save subsets
```{r}
#Run as a background job
for(i in 1:length(subset_list)){
  name = names(subset_list)[i]
  rna.cts = subset_list[[i]]
  so_subset <- subset_process_RNA_ATAC(so, rna.cts)
  p1 <- DimPlot(so_subset, reduction = "umap.rna", group.by = "timepoint_group", cols = colors_timepoints)
  p2 <- DimPlot(so_subset, reduction = "umap.rna", group.by = "annotation_broad", cols = colors_broad, label = T)
  p3 <- DimPlot(so_subset, reduction = "umap.rna", group.by = "RNA_snn_res.0.5", label = T)

  
  p4 <- DimPlot(so_subset, reduction = "umap.atac", group.by = "timepoint_group", cols = colors_timepoints)
  p5 <- DimPlot(so_subset, reduction = "umap.atac", group.by = "annotation_broad", cols = colors_broad, label = T)
  p6 <- DimPlot(so_subset, reduction = "umap.atac", group.by = "ATAC_snn_res.0.5", label = T)

  p7 <- DimPlot(so_subset, reduction = "wnn.umap", group.by = "timepoint_group", cols = colors_timepoints)
  p8 <- DimPlot(so_subset, reduction = "wnn.umap", group.by = "annotation_broad", cols = colors_broad, label = T)
  
  SaveFigure((p1|p2|p3)/(p4|p5|p6)/(p7|p8) & NoLegend() & NoAxes(), paste0(name, "_subset_umaps"), width = 10, height = 10)
  saveRDS(so_subset, paste0(outs_dir, name, "_so.rds"))
}
```

#run monocle
```{r}
#function for running monocle
runMonocle <- function(so){
  DefaultAssay(so) <- "RNA"
  cds <- as.cell_data_set(so, default.reduction = "umap.rna")
  rowData(cds)$gene_short_name <- translate(row.names(rowData(cds)))
  cds <- preprocess_cds(cds) 
  cds <- cluster_cells(cds)
  cds <- learn_graph(cds)
  cds <- order_cells(cds)
  return(cds)
}

```

#save as adata for input into RNA-velocity and CellRank

```{r}
i = 9
for(i in 1:length(subset_list)){
  name = names(subset_list)[i]
  so <- readRDS(paste0(outs_dir, name, "_so.rds"))
  
  #save umap coordinates to metadata
  so$umap.rna_1 <- so@reductions$umap.rna@cell.embeddings[,1]
  so$umap.rna_2 <- so@reductions$umap.rna@cell.embeddings[,2]
  so$umap.atac_1 <- so@reductions$umap.atac@cell.embeddings[,1]
  so$umap.atac_2 <- so@reductions$umap.atac@cell.embeddings[,2]
  so$umap.wnn_1 <- so@reductions$wnn.umap@cell.embeddings[,1]
  so$umap.wnn_2 <- so@reductions$wnn.umap@cell.embeddings[,2]
  
  # write metadata to csv
  so$barcode <- colnames(so)
  write.csv(so@meta.data, file=paste0(outs_dir, name, '_metadata.csv'), quote=F, row.names=F)
  
  # write pca matrix to csv
  write.csv(so@reductions$pca@cell.embeddings, file=paste0(outs_dir, name, '_pca.csv'), quote=F, row.names=F)
  
  # extract neighborhood graph
  nn_idx <- so@neighbors$weighted.nn@nn.idx
  nn_dist <- so@neighbors$weighted.nn@nn.dist
  nn_cells <- so@neighbors$weighted.nn@cell.names
  
  # save neighborhood graph
  write.table(nn_idx, paste0(outs_dir, name, "_nn_idx.txt"), sep = ',', row.names = F, col.names = F, quote = F)
  write.table(nn_dist, paste0(outs_dir, name, "_nn_dist.txt"), sep = ',', row.names = F, col.names = F, quote = F)
  write.table(nn_cells, paste0(outs_dir, name, "_nn_cells.txt"), sep = ',', row.names = F, col.names = F, quote = F)
}
```

```{r}
so_blood <- so
```


```{r}
# renv::install("zqfang/MuDataSeurat", force = T)
DefaultAssay(so_blood) <- "RNA"
so_blood_diet = DietSeurat(
  so_blood,
  layers = c("counts", "data"),
  features = rownames(so_blood), # export all genes, not just top highly variable genes
  assays = "RNA",
  dimreducs = c("pca","umap.rna"),
  graphs = c("RNA_nn", "RNA_snn"), # to RNA_nn -> distances, RNA_snn -> connectivities
  misc = TRUE
)
so_blood_diet <- JoinLayers(so_blood_diet)
MuDataSeurat::WriteH5AD(so_blood_diet, paste0(outs_dir, "so_blood.h5ad"), assay="RNA")
```
```{r}
counts <- GetAssayData(so_blood, slot = "counts")
writeMM(counts, file = paste0(outs_dir, name, "_counts.mtx"))
write.csv(rownames(counts), paste0(outs_dir, name, "_genes.csv"), row.names = FALSE)
write.csv(colnames(counts), paste0(outs_dir, name, "_barcodes.csv"), row.names = FALSE)

# 2. Save metadata
metadata <- so_blood@meta.data
write.csv(metadata, paste0(outs_dir, name, "_metadata.csv"))

# 3. Save dimensionality reductions (e.g., PCA and UMAP)
# PCA
write.csv(Embeddings(so_blood[["pca"]]), file = paste0(outs_dir, name, "_pca.csv"))
# UMAP
write.csv(Embeddings(so_blood[["umap.rna"]]), file = paste0(outs_dir, name, "_umap.csv"))
```
